Add difficulty support

+++ abdukulov-talgat.MathGame/abdukulov-talgat.MathGame/Core/GameSession.cs
@@ -24,7 +24,7 @@ public class GameSession : IEnumerable<RoundBase>
         }
     }
 
-    public int GetScore()
+    public float GetScore()
     {
         return _roundsList.Sum(r => r.GetScore());
     }
